refactor: line, section, station 모두와 관련된 공통된 기능은 common 패키지로 이동
