+++ src/main/java/subway/Application.java
@@ -1,5 +1,7 @@
 package subway;
 
+import subway.common.MainController;
+
 public class Application {
     public static void main(String[] args) {
         operateSubwaySystem();
